+++ src/VirtoCommerce.ElasticSearch8x.Core/Services/IElasticSearchFiltersBuilder.cs
@@ -1,0 +1,13 @@
+using System.Collections.Generic;
+using Elastic.Clients.Elasticsearch;
+using Elastic.Clients.Elasticsearch.Mapping;
+using Elastic.Clients.Elasticsearch.QueryDsl;
+using VirtoCommerce.SearchModule.Core.Model;
+
+namespace VirtoCommerce.ElasticSearch8x.Core.Services
+{
+    public interface IElasticSearchFiltersBuilder
+    {
+        Query GetFilterQuery(IFilter filter, IDictionary<PropertyName, IProperty> availableFields);
+    }
+}
