+++ src/VirtoCommerce.ElasticSearch8x.Core/Services/IElasticSearchResponseBuilder.cs
@@ -1,0 +1,12 @@
+using Elastic.Clients.Elasticsearch;
+using VirtoCommerce.SearchModule.Core.Model;
+using VirtoCommerceSearchRequest = VirtoCommerce.SearchModule.Core.Model.SearchRequest;
+
+
+namespace VirtoCommerce.ElasticSearch8x.Core.Services
+{
+    public interface IElasticSearchResponseBuilder
+    {
+        SearchResponse ToSearchResponse(SearchResponse<SearchDocument> response, VirtoCommerceSearchRequest request);
+    }
+}
