+++ src/VirtoCommerce.ElasticSearch8x.Core/VirtoCommerce.ElasticSearch8x.Core.csproj
@@ -12,6 +12,8 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="All" />
     <PackageReference Include="VirtoCommerce.Platform.Core" Version="3.253.1" />
+    <PackageReference Include="VirtoCommerce.SearchModule.Core" Version="3.223.0" />
+    <PackageReference Include="Elastic.Clients.Elasticsearch" Version="8.1.1" />
   </ItemGroup>
 
 </Project>
