+++ src/VirtoCommerce.ElasticSearch8x.Data/Extensions/ElasticSearchNameExtensions.cs
@@ -1,11 +1,4 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
 using System.Globalization;
-using System.Linq;
-using System.Reflection;
-using Elastic.Clients.Elasticsearch;
-using VirtoCommerce.SearchModule.Core.Model;
 
 namespace VirtoCommerce.ElasticSearch8x.Data.Extensions
 {
@@ -20,92 +13,5 @@ namespace VirtoCommerce.ElasticSearch8x.Data.Extensions
         {
             return string.Format(CultureInfo.InvariantCulture, "{0}", value);
         }
-    }
-
-    public static class ElasticSearchGeoPointExtensions
-    {
-        public static object ToElasticValue(this GeoPoint point)
-        {
-            return point == null ? null : new { lat = point.Latitude, lon = point.Longitude };
-        }
-
-        public static GeoLocation ToGeoLocation(this GeoPoint point)
-        {
-            return point == null ? null : GeoLocation.LatitudeLongitude(new LatLonGeoLocation { Lat = point.Latitude, Lon = point.Longitude });
-        }
-    }
-
-    public static class PropertyExtensions
-    {
-        public static ISet<string> GetPropertyNames<T>(this object obj, int deep)
-        {
-            var result = new HashSet<string>();
-            GetPropertyNamesInner<T>(obj, deep, parentName: null, result: result);
-
-            return result;
-        }
-
-        private static void GetPropertyNamesInner<T>(this object obj, int deep, string parentName, ISet<string> result)
-        {
-            if (obj == null || deep <= 0)
-            {
-                return;
-            }
-
-            var type = obj.GetType();
-
-            var properties = type
-                .GetProperties(BindingFlags.Instance | BindingFlags.Public)
-                .Where(x =>
-                    x.CanRead &&
-                    x.CanWrite &&
-                    (x.PropertyType == typeof(T) || HasNestedProperties(x.PropertyType)));
-
-            foreach (var property in properties)
-            {
-                if (property.PropertyType == typeof(T))
-                {
-                    var propertyName = GetFullName(parentName, property.Name);
-                    result.Add(propertyName);
-                    continue;
-                }
-
-                var value = property.GetValue(obj, null);
-
-                if (value is IEnumerable enumerable)
-                {
-                    var newParentName = GetFullName(parentName, property.Name);
-                    foreach (var item in enumerable)
-                    {
-                        GetPropertyNamesInner<T>(item, deep - 1, newParentName, result);
-                    }
-                }
-                else if (value != null && property.PropertyType.Assembly == type.Assembly)
-                {
-                    var newParentName = GetFullName(parentName, property.Name);
-                    GetPropertyNamesInner<T>(value, deep - 1, newParentName, result);
-                }
-            }
-        }
-
-        private static bool HasNestedProperties(Type type)
-        {
-            return
-                !type.IsPrimitive &&
-                type != typeof(string) &&
-                !type.IsAssignableTo(typeof(IEnumerable<string>));
-        }
-
-        private static string GetFullName(string parentName, string name)
-        {
-            return string.IsNullOrEmpty(parentName)
-                ? name.ToCamelCase()
-                : $"{parentName}.{name.ToCamelCase()}";
-        }
-
-        private static string ToCamelCase(this string str) =>
-            string.IsNullOrEmpty(str) || str.Length < 2
-                ? str
-                : char.ToLowerInvariant(str[0]) + str[1..];
     }
 }
