+++ src/VirtoCommerce.ElasticSearch8x.Data/VirtoCommerce.ElasticSearch8x.Data.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
@@ -17,8 +17,15 @@
     <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="All" />
     <PackageReference Include="VirtoCommerce.Platform.Data" Version="3.253.1" />
     <PackageReference Include="VirtoCommerce.SearchModule.Core" Version="3.223.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
     <PackageReference Include="Elastic.Clients.Elasticsearch" Version="8.1.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Models\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\VirtoCommerce.ElasticSearch8x.Core\VirtoCommerce.ElasticSearch8x.Core.csproj" />
   </ItemGroup>
 
 </Project>
