+++ src/VirtoCommerce.ElasticSearch8x.Web/Module.cs
@@ -1,10 +1,13 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using VirtoCommerce.ElasticSearch8x.Data.Models;
+using VirtoCommerce.ElasticSearch8x.Core;
+using VirtoCommerce.ElasticSearch8x.Core.Models;
+using VirtoCommerce.ElasticSearch8x.Core.Services;
 using VirtoCommerce.ElasticSearch8x.Data.Services;
 using VirtoCommerce.Platform.Core.Common;
 using VirtoCommerce.Platform.Core.Modularity;
+using VirtoCommerce.Platform.Core.Settings;
 using VirtoCommerce.SearchModule.Core.Services;
 
 namespace VirtoCommerce.ElasticSearch8x.Web;
@@ -30,14 +33,18 @@ public class Module : IModule, IHasConfiguration
             serviceCollection.Configure<ElasticSearch8xOptions>(Configuration.GetSection("Search:ElasticSearch8x"));
             serviceCollection.AddSingleton<ISearchProvider, ElasticSearch8xProvider>();
 
-            serviceCollection.AddSingleton<SearchRequestBuilder>();
-            serviceCollection.AddSingleton<SearchResponseBuilder>();
+            serviceCollection.AddSingleton<IElasticSearchFiltersBuilder, ElasticSearchFiltersBuilder>();
+            serviceCollection.AddSingleton<IElasticSearchAggregationsBuilder, ElasticSearchAggregationsBuilder>();
+
+            serviceCollection.AddSingleton<IElasticSearchRequestBuilder, ElasticSearchRequestBuilder>();
+            serviceCollection.AddSingleton<IElasticSearchResponseBuilder, ElasticSearchResponseBuilder>();
         }
     }
 
     public void PostInitialize(IApplicationBuilder appBuilder)
     {
-
+        var settingsRegistrar = appBuilder.ApplicationServices.GetRequiredService<ISettingsRegistrar>();
+        settingsRegistrar.RegisterSettings(ModuleConstants.Settings.AllSettings, ModuleInfo.Id);
     }
 
     public void Uninstall()
